third_party/WebKit/LayoutTests/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-ax-value-changed-notification.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<input id="test" type="datetime-local" value="2012-10-09T12:34:56">
<script>
description('This test checks value changed accessibility notifications.');

var testInput = document.getElementById('test');

if (!window.accessibilityController || !window.eventSender)
    debug('Please run inside DRT or WRT.');
else {
    accessibilityController.addNotificationListener(function (element, notification) {
        if (notification == 'Focus' || notification == 'ValueChanged') {
            debug(notification + ' ' + element.description + '=' + element.valueDescription);
        }
    });

    testInput.focus();
    eventSender.keyDown('\t');
    eventSender.keyDown('4');
    eventSender.keyDown('ArrowUp');
    eventSender.keyDown('\t');
    eventSender.keyDown('2');
    eventSender.keyDown('ArrowDown');

    window.jsTestIsAsync = true;

    window.setTimeout(function() {
        debug('');
        testInput.parentNode.removeChild(testInput);
        finishJSTest();
    }, 1);
}
</script>
</body>
</html>
