third_party/WebKit/LayoutTests/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-ax-value-changed-notification-expected.txt

This test checks value changed accessibility notifications.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Focus Month=AXValueDescription: 10
Focus Day=AXValueDescription: 09
Focus Year=AXValueDescription: 2012
Focus Hours=AXValueDescription: 12
Focus Minutes=AXValueDescription: 34
ValueChanged Day=AXValueDescription: 04
ValueChanged Day=AXValueDescription: 04
ValueChanged Year=AXValueDescription: 2013
ValueChanged Year=AXValueDescription: 2013
ValueChanged Hours=AXValueDescription: 02
ValueChanged Hours=AXValueDescription: 02
ValueChanged Minutes=AXValueDescription: 33
ValueChanged Minutes=AXValueDescription: 33

PASS successfullyParsed is true

TEST COMPLETE
